Switch all to user_id

+++ backend/app/web/api/briefcase/views.py
@@ -3,19 +3,18 @@ from typing import List
 
 from app.db.dao.briefcases import BriefcaseDAO
 from app.db.models.briefcase import (
-    BriefcaseItemModel,
     BriefcaseModel,
     BriefcaseRegistryModel,
 )
 from app.web.api.briefcase.scheme import (
     BriefcaseDTO,
     BriefcaseInputDTO,
-    BriefcaseItemDTO,
-    BriefcaseItemInputDTO,
     BriefcaseRegistryDTO,
     BriefcaseRegistryInputDTO,
 )
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, Depends, HTTPException
+
+from app.web.deps import CurrentUser, check_owner_or_superuser
 
 router = APIRouter()
 
@@ -23,122 +22,92 @@ router = APIRouter()
 @router.get("/{briefcase_id}", response_model=BriefcaseDTO)
 async def get_briefcase_model(
     briefcase_id: int,
+    current_user: CurrentUser,
     dao: BriefcaseDAO = Depends(),
 ) -> BriefcaseModel:
-    return await dao.get_briefcase_model(briefcase_id)
+    exist_briefcase = await dao.get_briefcase_model(briefcase_id)
+    await check_owner_or_superuser(exist_briefcase.user_id, current_user)
+
+    return exist_briefcase
 
 
 @router.get("/", response_model=List[BriefcaseDTO])
 async def get_briefcase_models(
+    current_user: CurrentUser,
     limit: int = 100,
     offset: int = 0,
     dao: BriefcaseDAO = Depends(),
 ) -> List[BriefcaseModel]:
-    return await dao.get_all_briefcases(limit=limit, offset=offset)
+    return await dao.get_all_briefcases(user_id=current_user.id, limit=limit, offset=offset)
 
 
 @router.post("/")
 async def create_briefcase_model(
     new_briefcase_object: BriefcaseInputDTO,
+    current_user: CurrentUser,
     dao: BriefcaseDAO = Depends(),
 ) -> None:
-    await dao.create_briefcase_model(fill_up=new_briefcase_object.fill_up)
+    await dao.create_briefcase_model(user_id=current_user.id, fill_up=new_briefcase_object.fill_up)
 
 
 @router.put("/{briefcase_id}")
 async def update_briefcase_model(
     briefcase_id: int,
     updated_briefcase: BriefcaseInputDTO,
-    briefcase_dao: BriefcaseDAO = Depends(),
+    current_user: CurrentUser,
+    dao: BriefcaseDAO = Depends(),
 ) -> None:
-    await briefcase_dao.update_briefcase_model(briefcase_id, updated_briefcase.fill_up)
+    exist_briefcase = await dao.get_briefcase_model(briefcase_id)
+    await check_owner_or_superuser(exist_briefcase.user_id, current_user)
+
+    await dao.update_briefcase_model(briefcase_id, updated_briefcase.fill_up)
 
 
 @router.delete("/{briefcase_id}", status_code=204)
 async def delete_briefcase_model(
     briefcase_id: int,
-    briefcase_dao: BriefcaseDAO = Depends(),
-) -> None:
-    await briefcase_dao.delete_briefcase_model(briefcase_id)
-
-
-@router.get("/items/{item_id}", response_model=BriefcaseItemDTO)
-async def get_briefcase_item(
-    item_id: int,
-    dao: BriefcaseDAO = Depends(),
-) -> BriefcaseItemModel:
-    return await dao.get_briefcase_item_model(item_id)
-
-
-@router.get("/{briefcase_id}/items/", response_model=List[BriefcaseItemDTO])
-async def get_briefcase_items(
-    briefcase_id: int,
-    dao: BriefcaseDAO = Depends(),
-) -> List[BriefcaseItemModel]:
-    return await dao.get_all_briefcase_items()
-
-
-@router.post("/{briefcase_id}/items/")
-async def create_briefcase_item_model(
-    briefcase_id: int,
-    new_item: BriefcaseItemInputDTO,  # Assuming you have a DTO for creating items
+    current_user: CurrentUser,
     dao: BriefcaseDAO = Depends(),
 ) -> None:
-    await dao.create_briefcase_item_model(
-        count=new_item.count,
-        dividends=new_item.dividends,
-        company_id=new_item.company.id,
-        strategy_id=new_item.strategy.id if new_item.strategy else None,
-        briefcase_id=briefcase_id,
-    )
+    exist_briefcase = await dao.get_briefcase_model(briefcase_id)
+    await check_owner_or_superuser(exist_briefcase.user_id, current_user)
+
+    await dao.delete_briefcase_model(briefcase_id)
 
 
-@router.put("/items/{item_id}")
-async def update_briefcase_item(
-    item_id: int,
-    updated_item: BriefcaseItemInputDTO,  # Assuming DTO for updating items
-    dao: BriefcaseDAO = Depends(),
-) -> None:
-    # Similar to creation, you'll need to pass the updated attributes to
-    # update_briefcase_item_model
-    await dao.update_briefcase_item_model(
-        briefcase_item_id=item_id,
-        count=updated_item.count,
-        dividends=updated_item.dividends,
-    )
-
-
-@router.delete("/items/{item_id}", status_code=204)
-async def delete_briefcase_item(
-    item_id: int,
-    dao: BriefcaseDAO = Depends(),
-) -> None:
-    """
-    Delete a specific item from a briefcase.
-
-    :param item_id: ID of the BriefcaseItemModel to delete.
-    :param dao: DAO for Briefcase models.
-    """
-    await dao.delete_briefcase_item_model(item_id)
-
-
-@router.get("/registry/{item_id}", response_model=BriefcaseRegistryDTO)
+@router.get("/{briefcase_id}/registry/{item_id}", response_model=BriefcaseRegistryDTO)
 async def get_briefcase_registry(
+    briefcase_id: int,
     item_id: int,
+    current_user: CurrentUser,
     dao: BriefcaseDAO = Depends(),
 ) -> BriefcaseRegistryModel:
-    return await dao.get_briefcase_registry_model(item_id)
+    exist_briefcase = await dao.get_briefcase_model(briefcase_id)
+    await check_owner_or_superuser(exist_briefcase.user_id, current_user)
+
+    exist_item = await dao.get_briefcase_registry_model(item_id)
+    if exist_item.briefcase.id != briefcase_id:
+        raise HTTPException(
+            status_code=400,
+            detail="The item_id doesn't match briefcase",
+        )
+
+    return exist_item
 
 
 @router.get("/{briefcase_id}/registry/", response_model=List[BriefcaseRegistryDTO])
 async def get_briefcase_registries(  # noqa: WPS211
     briefcase_id: int,
+    current_user: CurrentUser,
     dao: BriefcaseDAO = Depends(),
     limit: int = 100000,
     offset: int = 0,
     date_from: str = None,
     date_to: str = None,
 ) -> List[BriefcaseRegistryModel]:
+    exist_briefcase = await dao.get_briefcase_model(briefcase_id)
+    await check_owner_or_superuser(exist_briefcase.user_id, current_user)
+
     return await dao.get_all_briefcase_registry(
         briefcase_id=briefcase_id,
         limit=limit,
@@ -151,9 +120,13 @@ async def get_briefcase_registries(  # noqa: WPS211
 @router.post("/{briefcase_id}/registry/")
 async def create_briefcase_registry_model(
     briefcase_id: int,
+    current_user: CurrentUser,
     new_item: BriefcaseRegistryInputDTO,  # Assuming you have a DTO for creating items
     dao: BriefcaseDAO = Depends(),
 ) -> None:
+    exist_briefcase = await dao.get_briefcase_model(briefcase_id)
+    await check_owner_or_superuser(exist_briefcase.user_id, current_user)
+
     await dao.create_briefcase_registry_model(
         count=new_item.count,
         amount=new_item.amount,
@@ -167,21 +140,45 @@ async def create_briefcase_registry_model(
     )
 
 
-@router.put("/registry/{item_id}")
+@router.put("/{briefcase_id}/registry/{item_id}")
 async def update_briefcase_registry(
+    briefcase_id: int,
     item_id: int,
+    current_user: CurrentUser,
     updated_item: BriefcaseRegistryInputDTO,
     dao: BriefcaseDAO = Depends(),
 ) -> None:
+    exist_briefcase = await dao.get_briefcase_model(briefcase_id)
+    await check_owner_or_superuser(exist_briefcase.user_id, current_user)
+
+    exist_item = await dao.get_briefcase_registry_model(item_id)
+    if exist_item.briefcase.id != briefcase_id:
+        raise HTTPException(
+            status_code=400,
+            detail="The item_id doesn't match briefcase",
+        )
+
     await dao.update_briefcase_registry_model(
         registry_id=item_id,
         updated_fields=updated_item.model_dump(),
     )
 
 
-@router.delete("/registry/{item_id}", status_code=204)
+@router.delete("/{briefcase_id}/registry/{item_id}", status_code=204)
 async def delete_briefcase_registry(
+    briefcase_id: int,
     item_id: int,
+    current_user: CurrentUser,
     dao: BriefcaseDAO = Depends(),
 ) -> None:
+    exist_briefcase = await dao.get_briefcase_model(briefcase_id)
+    await check_owner_or_superuser(exist_briefcase.user_id, current_user)
+
+    exist_item = await dao.get_briefcase_registry_model(item_id)
+    if exist_item.briefcase.id != briefcase_id:
+        raise HTTPException(
+            status_code=400,
+            detail="The item_id doesn't match briefcase",
+        )
+
     await dao.delete_briefcase_registry_model(item_id)
